Put the logic into super class heap
there are 2 specilized classes called MinHeap and MaxHeap, both inheriting from Heap class.
the only logical difference between how Max/Min heap operate is the way they prioritize things.
there is a method called hasHigherPriority, which uses < operator for minHeap and > operator for MaxHeap to determine the priprity of the items. if a heap of custom object is created, those operators must be overloaded.

The reason for using these 2 operators was to avoid unnecessary data movement when not necessary. (we could have implemented with only one operator for both sub classes)

Signed-off-by: Hassan Attar <hattar00@mylangara.ca>